pyproject.toml
[project]
name = "vla-driven-manipulation"
version = "0.1.0"
description = "Phase 1 GRPO reward validation on classic RL benchmarks."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "gymnasium>=0.29",
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
  "scipy>=1.13",
  "tensorboard>=2.16",
  "torch>=2.2",
]

[project.optional-dependencies]
mujoco = [
  "gymnasium[mujoco]>=0.29",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

